Name check
Check numbers in name method checks if there are numbers in a name.
Join middle names takes a list and saves the first value and the last value into a new list while the values are in between an empty string.
All of the middle names get joined together but separated by - as a delimiter.
Create name method checks if the name is valid and sends  back 3 values(firstname, middlename, lastname).