Client init method logic base
Method checks if method recieves user_data or user_id. If the method recieves only a list of values, it initiates a guest client class.
If it recieves only the user_id then it initialises the user from the database.
If it recieves both, it initialises a new registered user for the session.