Existing client init method logic
Load user data from db takes the entire database and runs through it. If user id is found in the current list, it returns that list.
Registered user initiation saves the loaded data into a variable. It then takes the data inside the variable list and attempts to convert all of the variables into lists. If the item is not a list, it skips the item. Then it splits the clients name and checks if the clients name list length is larger than 2. If it is, it assigns each name to a different variable corresponding to its position. If the length is 2, it assigns the first and last name while making the middle name into an empty string.
It then creates a registered client class with that data, sets the user id and sets the completed and pending orders inside a class method the user has.