User removal
Using panda package, reads the entirety of the database into the df(dataframe) variable and iterates through the panda indexes using .iloc method. If the currently indexed data in the client_id column is equal to the methods client_id then it drops the row from the database and writes it back into the database file and stops iterating through the file.