Database fix
Imported OS module and defined an absolute path for the database read/write methods. Removed the client_db gloabl variable and made every method which tries to read the database load the database into a variable. Added an empty line at the end of main for a better viewing experience